Product Backlog Item 166292: Rename prepare, cleanup & delete methods
