+++ package/cloudshell/tests/test_commands/test_command_orchestrator.py
@@ -88,9 +88,9 @@ class TestCommandOrchestrator(TestCase):
         # assert
         self.assertTrue(self.command_orchestrator.command_wrapper.execute_command_with_connection.called)
 
-    def test_destroy_vm_only(self):
+    def test_DeleteInstance(self):
         # act
-        self.command_orchestrator.destroy_vm_only(self.context, self.ports)
+        self.command_orchestrator.DeleteInstance(self.context, self.ports)
         # assert
         self.assertTrue(self.command_orchestrator.command_wrapper.execute_command_with_connection.called)
